Refactor KubernetesPipelineTest
* Pull TestName rule in AbstractKubernetesPipelineTest
* Factor boilerplate in a @Before method
* Camelcase method name generates job with space
* Move test cases with overriden namespace to a separate test class
* Some fixes related to space use
