+++ src/test/java/org/csanchez/jenkins/plugins/kubernetes/pipeline/ContainerLogStepTest.java
@@ -28,14 +28,6 @@ import static org.junit.Assert.assertNotNull;
 
 public class ContainerLogStepTest extends AbstractKubernetesPipelineTest {
 
-    @Rule
-    public TestName name = new TestName();
-
-    @Override
-    protected TestName getTestName() {
-        return name;
-    }
-
     @Issue("JENKINS-46085")
     @Test
     public void simple() throws Exception {
