+++ src/test/resources/org/csanchez/jenkins/plugins/kubernetes/pipeline/runInPod.groovy
@@ -10,7 +10,7 @@ podTemplate(label: 'runInPod', containers: [
             ##
             echo "script file: \$(find ../../.. -iname script.sh))"
             echo "script file contents: \$(find ../../.. -iname script.sh -exec cat {} \\;)"
-            test -n "\$(cat \$(find ../../.. -iname script.sh))"
+            test -n "\$(cat \"\$(find ../../.. -iname script.sh)\")"
           """
         }
       }
