+++ src/test/resources/org/csanchez/jenkins/plugins/kubernetes/pipeline/runInPodFromYaml.groovy
@@ -25,7 +25,7 @@ spec:
             ##
             echo "script file: \$(find ../../.. -iname script.sh))"
             echo "script file contents: \$(find ../../.. -iname script.sh -exec cat {} \\;)"
-            test -n "\$(cat \$(find ../../.. -iname script.sh))"
+            test -n "\$(cat \"\$(find ../../.. -iname script.sh)\")"
           """
         }
       }
