+++ src/test/resources/org/csanchez/jenkins/plugins/kubernetes/pipeline/runWithEnvVariablesInContext.groovy
