+++ src/test/resources/org/csanchez/jenkins/plugins/kubernetes/pipeline/supportComputerEnvVars.groovy
